refactor: move dead code from polynomial/mod.rs to test files
Move test-only functions from the polynomial module to test files where
they are used:
- evaluate_slice, differentiate, ruffini_division, xgcd, div_with_ref,
  print_as_sage_poly, compose, and Div trait impl moved to polynomial_tests.rs
- evaluate_slice helper added to 6 FFT test files

This cleans up the public API by removing functions only used in tests.
