+++ crypto/math/src/tests/fft_friendly_quartic_babybear_tests.rs
@@ -137,6 +137,15 @@ mod test_babybear_31_fft {
     use crate::polynomial::Polynomial;
     use proptest::{collection, prelude::*, std_facade::Vec};
 
+    /// Evaluates a polynomial at a slice of points
+    #[cfg(not(feature = "cuda"))]
+    fn evaluate_slice<F: IsFFTField>(
+        poly: &Polynomial<FieldElement<F>>,
+        input: &[FieldElement<F>],
+    ) -> Vec<FieldElement<F>> {
+        input.iter().map(|x| poly.evaluate(x)).collect()
+    }
+
     #[cfg(not(feature = "cuda"))]
     fn gen_fft_and_naive_evaluation<F: IsFFTField>(
         poly: Polynomial<FieldElement<F>>,
@@ -147,7 +156,7 @@ mod test_babybear_31_fft {
             get_powers_of_primitive_root(order.into(), len, RootsConfig::Natural).unwrap();
 
         let fft_eval = Polynomial::evaluate_fft::<F>(&poly, 1, None).unwrap();
-        let naive_eval = poly.evaluate_slice(&twiddles);
+        let naive_eval = evaluate_slice(&poly, &twiddles);
 
         (fft_eval, naive_eval)
     }
@@ -165,7 +174,7 @@ mod test_babybear_31_fft {
 
         let fft_eval =
             Polynomial::evaluate_offset_fft::<F>(&poly, blowup_factor, None, &offset).unwrap();
-        let naive_eval = poly.evaluate_slice(&twiddles);
+        let naive_eval = evaluate_slice(&poly, &twiddles);
 
         (fft_eval, naive_eval)
     }
