+++ crypto/math/src/tests/fft_tests.rs
@@ -59,6 +59,14 @@ mod fft_polynomial_tests {
     use crate::polynomial::Polynomial;
     use proptest::{collection, prelude::*};
 
+    /// Evaluates a polynomial at a slice of points
+    fn evaluate_slice<F: IsFFTField>(
+        poly: &Polynomial<FieldElement<F>>,
+        input: &[FieldElement<F>],
+    ) -> Vec<FieldElement<F>> {
+        input.iter().map(|x| poly.evaluate(x)).collect()
+    }
+
     fn gen_fft_and_naive_evaluation<F: IsFFTField>(
         poly: Polynomial<FieldElement<F>>,
     ) -> (Vec<FieldElement<F>>, Vec<FieldElement<F>>) {
@@ -68,7 +76,7 @@ mod fft_polynomial_tests {
             get_powers_of_primitive_root(order.into(), len, RootsConfig::Natural).unwrap();
 
         let fft_eval = Polynomial::evaluate_fft::<F>(&poly, 1, None).unwrap();
-        let naive_eval = poly.evaluate_slice(&twiddles);
+        let naive_eval = evaluate_slice(&poly, &twiddles);
 
         (fft_eval, naive_eval)
     }
@@ -85,7 +93,7 @@ mod fft_polynomial_tests {
 
         let fft_eval =
             Polynomial::evaluate_offset_fft::<F>(&poly, blowup_factor, None, &offset).unwrap();
-        let naive_eval = poly.evaluate_slice(&twiddles);
+        let naive_eval = evaluate_slice(&poly, &twiddles);
 
         (fft_eval, naive_eval)
     }
